+++ changelog.md
@@ -1,2 +1,2 @@
 ### Rechiseled 1.1.6
-- Initial release of Rechiseled for Minecraft 1.21.9 & 1.21.10
+- Initial release of Rechiseled for Minecraft 1.21.11